Update start.py
Most spiciest change yet. Called from the command terminal together with a .yml file as an argument, the initialiser will only see a string that (hopefully) contains a path. But it wants a dictionary. Now it can load it's own dictionary when provided with a path or it can be loaded with a dictionary directly.